test: MenuProducts의 품목들에 대한 모든 값에 대한 합을 구하는 것에 대한 테스트 코드 추가

+++ src/main/java/kitchenpos/menu/domain/MenuProducts.java
@@ -20,7 +20,7 @@ public class MenuProducts {
 
 		for (final MenuProduct menuProduct : menuProducts) {
 			Product product = menuProduct.getProduct();
-			sum = sum.add(product.getPrice().multiply(BigDecimal.valueOf(menuProduct.getQuantity())));
+			sum = sum.add(product.getPrice().multiply(menuProduct.getQuantity()));
 		}
 
 		return sum;
